refactor: only expose two APIs for setting inputs
one without triggering cd, the other with
