+++ src/app/resume-page/chip/chip.component.spec.ts
@@ -3,7 +3,10 @@ import { componentTestSetup } from '@/test/helpers/component-test-setup'
 import { ComponentFixture } from '@angular/core/testing'
 import { first, Subscription } from 'rxjs'
 import { shouldProjectContent } from '@/test/helpers/component-testers'
-import { setFixtureInput } from '@/test/helpers/set-input'
+import {
+  setFixtureInputs,
+  setFixtureInputsAndDetectChanges,
+} from '@/test/helpers/set-fixture-inputs'
 
 describe('ChipComponent', () => {
   it('should create', () => {
@@ -18,7 +21,7 @@ describe('ChipComponent', () => {
   describe('when selected attribute is false', () => {
     it('should not add the selected class', () => {
       const [fixture] = componentTestSetup(ChipComponent)
-      setFixtureInput(fixture, 'isSelected', false)
+      setFixtureInputsAndDetectChanges(fixture, { isSelected: false })
 
       expect(fixture.debugElement.classes['selected']).toBeFalsy()
     })
@@ -27,7 +30,7 @@ describe('ChipComponent', () => {
   describe('when selected attribute is true', () => {
     it('should add the selected class', () => {
       const [fixture] = componentTestSetup(ChipComponent)
-      setFixtureInput(fixture, 'isSelected', true)
+      setFixtureInputsAndDetectChanges(fixture, { isSelected: true })
 
       expect(fixture.debugElement.classes['selected']).toBeTrue()
     })
@@ -64,7 +67,7 @@ describe('ChipComponent', () => {
 
     beforeEach(() => {
       ;[fixture, component] = componentTestSetup(ChipComponent)
-      setFixtureInput(fixture, 'isSelected', false, { noChangeDetection: true })
+      setFixtureInputs(fixture, { isSelected: false })
       subscription = component.isSelectedChange
         .pipe(first())
         .subscribe((isSelectedReceived) => (isSelected = isSelectedReceived))
