+++ src/app/resume-page/date-range/date-range.component.spec.ts
@@ -4,7 +4,7 @@ import { DateRangeComponent } from './date-range.component'
 import { By } from '@angular/platform-browser'
 import { DateRange } from './date-range'
 import { textContent } from '@/test/helpers/text-content'
-import { setFixtureInput } from '@/test/helpers/set-input'
+import { setFixtureInputsAndDetectChanges } from '@/test/helpers/set-fixture-inputs'
 
 describe('DateRangeComponent', () => {
   let component: DateRangeComponent
@@ -21,7 +21,7 @@ describe('DateRangeComponent', () => {
   })
 
   function setRangeAndDetectChanges(range: DateRange) {
-    setFixtureInput(fixture, 'range', range)
+    setFixtureInputsAndDetectChanges(fixture, { range })
   }
 
   it('should display start date formatted as month abbreviation and year', () => {
