+++ src/app/resume-page/languages-section/language-item/language-item.component.spec.ts
@@ -15,7 +15,7 @@ import { CardHeaderTextsComponent } from '../../card/card-header/card-header-tex
 import { CardHeaderSubtitleComponent } from '../../card/card-header/card-header-subtitle/card-header-subtitle.component'
 import { TestIdDirective } from '@/common/test-id.directive'
 import { textContent } from '@/test/helpers/text-content'
-import { setFixtureInput } from '@/test/helpers/set-input'
+import { setFixtureInputsAndDetectChanges } from '@/test/helpers/set-fixture-inputs'
 
 describe('LanguageItemComponent', () => {
   let component: LanguageItemComponent
@@ -115,5 +115,7 @@ function setLanguageItem(
   fixture: ComponentFixture<LanguageItemComponent>,
   overrides?: Partial<LanguageItem>,
 ) {
-  setFixtureInput(fixture, 'item', makeLanguageItem(overrides))
+  setFixtureInputsAndDetectChanges(fixture, {
+    item: makeLanguageItem(overrides),
+  })
 }
